use 32bit to represent data

+++ pill/timedfifo.h
@@ -15,7 +15,7 @@
 #define TF_UNIT_TIME_S 60 
 #define TF_UNIT_TIME_MS 60000
 #define TF_BUFFER_SIZE (8 * 60)
-typedef int16_t tf_unit_t;  // Good job, this is a keen design!
+typedef int32_t tf_unit_t;  // Good job, this is a keen design!
 
 typedef struct{
     uint8_t version;
